Further optimisation
When a hidden fixed point (meaning an element which is the only element in its row/column/box that can take a certain value, or an element which can only take one value) is found, it is removed from the arrays of possible values for the other elements in its row/column/box. 
So far, I have only done this to find hidden fixed points in the rows.