Sudoku Solver - Brute Force Method
This solution is a brute force method - it goes through every element in the grid, going through the values 1-9, and when it finds a value that is valid according to the standard rules of Sudoku (no repeats on a row/column/box), it goes to the next box and repeats. If an element gets to the number 9 without finding a valid value, it returns to the previous box and increments that element to the next valid value, and so on, until a complete solution is found, or it is found that there is no solution.